moved domain classes
