+++ app/src/main/java/be/nilsberghs/galileoproject/ScoreViewModel.kt
@@ -2,28 +2,24 @@ package be.nilsberghs.galileoproject
 
 import android.app.Application
 import android.content.Context
-import android.content.SharedPreferences
 import android.util.Log
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import be.nilsberghs.galileoproject.data.AppDatabase
-import be.nilsberghs.galileoproject.data.Game
-import be.nilsberghs.galileoproject.data.GameHistory
-import be.nilsberghs.galileoproject.data.Player
-import be.nilsberghs.galileoproject.data.ScoreEntry
-import be.nilsberghs.galileoproject.data.ScoreWithPlayer
+import be.nilsberghs.galileoproject.domain.Game
+import be.nilsberghs.galileoproject.domain.GameHistory
+import be.nilsberghs.galileoproject.domain.Player
+import be.nilsberghs.galileoproject.domain.ScoreEntry
+import be.nilsberghs.galileoproject.domain.ScoreWithPlayer
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import androidx.core.content.edit
 import be.nilsberghs.galileoproject.util.AddPlayerResult
