+++ app/src/main/java/be/nilsberghs/galileoproject/data/PlayerDao.kt
@@ -5,6 +5,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
+import be.nilsberghs.galileoproject.domain.Player
 import kotlinx.coroutines.flow.Flow
 
 @Dao
